Move `IntegrationTestCase` and `NodeTestCase` from `/src` to `/tests`
