+++ extra/html-extra/tests/IntegrationTest.php
@@ -12,7 +12,7 @@
 namespace Twig\Extra\Html\Tests;
 
 use Twig\Extra\Html\HtmlExtension;
-use Twig\Test\IntegrationTestCase;
+use Twig\Tests\IntegrationTestCase;
 
 class IntegrationTest extends IntegrationTestCase
 {
