+++ tests/Node/DoTest.php
@@ -13,7 +13,7 @@ namespace Twig\Tests\Node;
 
 use Twig\Node\DoNode;
 use Twig\Node\Expression\ConstantExpression;
-use Twig\Test\NodeTestCase;
+use Twig\Tests\NodeTestCase;
 
 class DoTest extends NodeTestCase
 {
