+++ tests/Node/Expression/ArrayTest.php
@@ -13,7 +13,7 @@ namespace Twig\Tests\Node\Expression;
 
 use Twig\Node\Expression\ArrayExpression;
 use Twig\Node\Expression\ConstantExpression;
-use Twig\Test\NodeTestCase;
+use Twig\Tests\NodeTestCase;
 
 class ArrayTest extends NodeTestCase
 {
