+++ tests/Node/Expression/Binary/AndTest.php
@@ -13,7 +13,7 @@ namespace Twig\Tests\Node\Expression\Binary;
 
 use Twig\Node\Expression\Binary\AndBinary;
 use Twig\Node\Expression\ConstantExpression;
-use Twig\Test\NodeTestCase;
+use Twig\Tests\NodeTestCase;
 
 class AndTest extends NodeTestCase
 {
