+++ tests/Node/Expression/Binary/ConcatTest.php
@@ -13,7 +13,7 @@ namespace Twig\Tests\Node\Expression\Binary;
 
 use Twig\Node\Expression\Binary\ConcatBinary;
 use Twig\Node\Expression\ConstantExpression;
-use Twig\Test\NodeTestCase;
+use Twig\Tests\NodeTestCase;
 
 class ConcatTest extends NodeTestCase
 {
