+++ tests/Node/Expression/Binary/OrTest.php
@@ -13,7 +13,7 @@ namespace Twig\Tests\Node\Expression\Binary;
 
 use Twig\Node\Expression\Binary\OrBinary;
 use Twig\Node\Expression\ConstantExpression;
-use Twig\Test\NodeTestCase;
+use Twig\Tests\NodeTestCase;
 
 class OrTest extends NodeTestCase
 {
