+++ tests/Node/Expression/ConditionalTest.php
@@ -13,7 +13,7 @@ namespace Twig\Tests\Node\Expression;
 
 use Twig\Node\Expression\ConditionalExpression;
 use Twig\Node\Expression\ConstantExpression;
-use Twig\Test\NodeTestCase;
+use Twig\Tests\NodeTestCase;
 
 class ConditionalTest extends NodeTestCase
 {
