+++ tests/Node/Expression/NullCoalesceTest.php
@@ -14,7 +14,7 @@ namespace Twig\Tests\Node\Expression;
 use Twig\Node\Expression\ConstantExpression;
 use Twig\Node\Expression\NameExpression;
 use Twig\Node\Expression\NullCoalesceExpression;
-use Twig\Test\NodeTestCase;
+use Twig\Tests\NodeTestCase;
 
 class NullCoalesceTest extends NodeTestCase
 {
