+++ tests/Node/IncludeTest.php
@@ -15,7 +15,7 @@ use Twig\Node\Expression\ArrayExpression;
 use Twig\Node\Expression\ConditionalExpression;
 use Twig\Node\Expression\ConstantExpression;
 use Twig\Node\IncludeNode;
-use Twig\Test\NodeTestCase;
+use Twig\Tests\NodeTestCase;
 
 class IncludeTest extends NodeTestCase
 {
